version update

+++ src/test/java/org/opensrp/service/TaskServiceTest.java
@@ -189,7 +189,7 @@ public class TaskServiceTest {
 
 		ArgumentCaptor<Task> argumentCaptor = ArgumentCaptor.forClass(Task.class);
 		verify(taskRepository).update(argumentCaptor.capture());
-		assertTrue(argumentCaptor.getValue().getServerVersion().longValue()==now);
+		assertTrue(argumentCaptor.getValue().getServerVersion()==now);
 		assertEquals(argumentCaptor.getValue().getBusinessStatus(),"Not Sprayable");
 	}
